refactoring
refactored useApi hook
working
